@@ -2,9 +2,10 @@
   "name": "pixel-game-engine",
   "version": "0.1.0",
   "description": "Simple game engine inspired by OneLoneCoder",
-  "main": "build/index.js",
+  "main": "index.js",
+	"type": "module",
   "scripts": {
-    "start": "tsc index.ts",
+    "start": "tsc",
     "test": "jest",
     "test:watch": "jest --watch",
     "build:watch": "tsc -w"
@@ -14,11 +15,11 @@
   "devDependencies": {
     "@babel/preset-env": "^7.19.4",
     "@babel/preset-typescript": "^7.18.6",
-    "@types/jest": "^29.1.2",
+    "@types/jest": "^29.5.13",
     "@types/random-seed": "^0.3.3",
-    "jest": "^27.5.1",
-    "ts-jest": "^27.1.3",
-    "typescript": "^4.5.5"
+    "jest": "^29.7.0",
+    "ts-jest": "^29.2.5",
+    "typescript": "^5.6.3"
   },
   "dependencies": {
     "random-seed": "^0.3.0"
