@@ -1,22 +1,10 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.Random = exports.Vector = exports.Position = exports.Entity = exports.Grid = exports.Area = exports.COLORS = exports.Color = exports.COLOURS = exports.Colour = exports.PixelGameEngine = void 0;
-var PixelGameEngine_1 = require("./src/PixelGameEngine");
-Object.defineProperty(exports, "PixelGameEngine", { enumerable: true, get: function () { return PixelGameEngine_1.PixelGameEngine; } });
-var Colour_1 = require("./src/Colour");
-Object.defineProperty(exports, "Colour", { enumerable: true, get: function () { return Colour_1.Colour; } });
-Object.defineProperty(exports, "COLOURS", { enumerable: true, get: function () { return Colour_1.COLOURS; } });
-var Colour_2 = require("./src/Colour");
-Object.defineProperty(exports, "Color", { enumerable: true, get: function () { return Colour_2.Colour; } });
-Object.defineProperty(exports, "COLORS", { enumerable: true, get: function () { return Colour_2.COLORS; } });
-var Area_1 = require("./src/locationObjects/Area");
-Object.defineProperty(exports, "Area", { enumerable: true, get: function () { return Area_1.default; } });
-var Grid_1 = require("./src/Grid");
-Object.defineProperty(exports, "Grid", { enumerable: true, get: function () { return Grid_1.Grid; } });
-var Entity_1 = require("./src/locationObjects/Entity");
-Object.defineProperty(exports, "Entity", { enumerable: true, get: function () { return Entity_1.Entity; } });
-var Position_1 = require("./src/locationObjects/Position");
-Object.defineProperty(exports, "Position", { enumerable: true, get: function () { return Position_1.default; } });
-Object.defineProperty(exports, "Vector", { enumerable: true, get: function () { return Position_1.default; } });
-var Random_1 = require("./src/tools/Random");
-Object.defineProperty(exports, "Random", { enumerable: true, get: function () { return Random_1.Random; } });
+export { PixelGameEngine as PixelGameEngine } from './src/PixelGameEngine';
+export { default as Area } from './src/locationObjects/Area';
+export { default as collision } from './src/tools/collision';
+export { Colour, COLOURS } from "./src/Colour";
+export { Colour as Color, COLORS } from "./src/Colour";
+export { Entity } from './src/locationObjects/Entity';
+export { Grid } from "./src/Grid";
+export { Random } from './src/tools/Random';
+export { default as Position, default as Vector } from './src/locationObjects/Position';
+//# sourceMappingURL=index.js.map
