@@ -1,8 +1,9 @@
 export {PixelGameEngine as PixelGameEngine} from './src/PixelGameEngine';
+export {default as Area} from './src/locationObjects/Area';
+export {default as collision} from './src/tools/collision'
 export {Colour, COLOURS} from "./src/Colour";
 export {Colour as Color, COLORS} from "./src/Colour";
-export {default as Area} from './src/locationObjects/Area';
-export {Grid} from "./src/Grid";
 export {Entity} from './src/locationObjects/Entity'
-export {default as Position, default as Vector} from './src/locationObjects/Position';
+export {Grid} from "./src/Grid";
 export {Random} from './src/tools/Random';
+export {default as Position, default as Vector} from './src/locationObjects/Position';
